Move TimeDerivStrahlkorper tag
