+++ src/PointwiseFunctions/GeneralRelativity/Surfaces/Tags.hpp
@@ -14,7 +14,6 @@
 #include "DataStructures/Tensor/EagerMath/Magnitude.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
 #include "NumericalAlgorithms/SphericalHarmonics/Strahlkorper.hpp"
-#include "NumericalAlgorithms/SphericalHarmonics/StrahlkorperFunctions.hpp"
 #include "NumericalAlgorithms/SphericalHarmonics/TagsDeclarations.hpp"
 #include "NumericalAlgorithms/SphericalHarmonics/TagsTypeAliases.hpp"
 #include "PointwiseFunctions/GeneralRelativity/Surfaces/AreaElement.hpp"
@@ -34,36 +33,6 @@
 /// \ingroup SurfacesGroup
 /// Holds tags and ComputeItems associated with a `::Strahlkorper`.
 namespace StrahlkorperTags {
-/// Tag for holding the previously-found values of a Strahlkorper,
-/// which are saved for extrapolation for future initial guesses
-/// and for computing the time deriv of a Strahlkorper.
-template <typename Frame>
-struct PreviousStrahlkorpers : db::SimpleTag {
-  using type = std::deque<std::pair<double, ::Strahlkorper<Frame>>>;
-};
-
-/// @{
-/// Tag to compute the time derivative of the coefficients of a Strahlkorper
-/// from a number of previous Strahlkorpers.
-template <typename Frame>
-struct TimeDerivStrahlkorper : db::SimpleTag {
-  using type = ::Strahlkorper<Frame>;
-};
-
-template <typename Frame>
-struct TimeDerivStrahlkorperCompute : db::ComputeTag,
-                                      TimeDerivStrahlkorper<Frame> {
-  using base = TimeDerivStrahlkorper<Frame>;
-  using return_type = typename base::type;
-  static constexpr auto function = static_cast<void (*)(
-      gsl::not_null<::Strahlkorper<Frame>*>,
-      const std::deque<std::pair<double, ::Strahlkorper<Frame>>>&)>(
-      &StrahlkorperFunctions::time_deriv_of_strahlkorper<Frame>);
-
-  using argument_tags = tmpl::list<PreviousStrahlkorpers<Frame>>;
-};
-/// @}
-
 /// The OneOverOneFormMagnitude is the reciprocal of the magnitude of the
 /// one-form perpendicular to the horizon
 struct OneOverOneFormMagnitude : db::SimpleTag {
